refactor(section): section의 라이플 사이클에 따라 모든 지하철 구간 관련 파일을 line 패키지에 종속, 이관시킵니다.
ref: https://github.com/next-step/atdd-subway-map/pull/710#discussion_r1097450305
