+++ src/main/java/subway/line/application/LineService.java
@@ -8,6 +8,8 @@ import subway.line.domain.Line;
 import subway.line.domain.LineCommandRepository;
 import subway.line.domain.LineQueryRepository;
 import subway.line.exception.LineNotFoundException;
+import subway.line.application.dto.request.SectionCreateRequest;
+import subway.line.domain.Section;
 import subway.station.application.StationService;
 import subway.station.domain.Station;
 
@@ -59,6 +61,25 @@ public class LineService {
         lineCommandRepository.deleteById(lineId);
     }
 
+    @Transactional
+    public Long saveSection(final Long lineId, final SectionCreateRequest sectionCreateRequest) {
+        Line findLine = findLineById(lineId);
+        Station upStation = stationService.findStationById(sectionCreateRequest.getUpStationId());
+        Station downStation = stationService.findStationById(sectionCreateRequest.getDownStationId());
+        Section section = Section.createSection(findLine, upStation, downStation, sectionCreateRequest.getDistance());
+
+        findLine.getSections().addSection(section);
+
+        return section.getId();
+    }
+
+    @Transactional
+    public void deleteSection(final Long lineId, final Long stationId) {
+        Line findLine = findLineById(lineId);
+        Station findStation = stationService.findStationById(stationId);
+
+        findLine.getSections().remove(findStation);
+    }
 
     private Line createLine(final LineCreateRequest lineCreateRequest) {
         Station upStation = stationService.findStationById(lineCreateRequest.getUpStationId());
